Update usecase naming
