+++ domain/src/main/java/com/dsm/dms/domain/DomainExtensions.kt
@@ -10,6 +10,12 @@ fun Completable.toSingleResult(
     handler: ErrorHandler
 ): Single<Result<Unit>> = this
     .toSingle { Unit }.toResult(handler)
+    .map {
+        when(it) {
+            is Result.Success -> it
+            is Result.Error -> Result.Error(Unit, it.message)
+        }
+    }
 
 fun <T> Completable.toSingleResult(
     handler: ErrorHandler,
@@ -19,6 +25,12 @@ fun <T> Completable.toSingleResult(
     .toSingle {
         localFun(data)
     }.toResult(handler)
+    .map {
+        when(it) {
+            is Result.Success -> it
+            is Result.Error -> Result.Error(Unit, it.message)
+        }
+    }
 
 fun <T> Single<T>.toResult(
     handler: ErrorHandler,
