+++ domain/src/main/java/com/dsm/dms/domain/usecase/ApplyExtensionUseCase.kt
@@ -2,17 +2,17 @@ package com.dsm.dms.domain.usecase
 
 import com.dsm.dms.domain.base.Result
 import com.dsm.dms.domain.base.UseCase
+import com.dsm.dms.domain.entity.ExtensionInfo
 import com.dsm.dms.domain.service.ExtensionService
 import io.reactivex.Single
 import io.reactivex.disposables.CompositeDisposable
 
-
-class DeleteExtensionInfoUseCase(
+class ApplyExtensionUseCase(
     private val service: ExtensionService,
     composite: CompositeDisposable
-): UseCase<Int, Result<Unit>>(composite) {
+): UseCase<ExtensionInfo, Result<Unit>>(composite) {
 
-    override fun create(time: Int): Single<Result<Unit>> =
-        service.deleteExtensionInfo(time)
+    override fun create(extensionInfo: ExtensionInfo): Single<Result<Unit>> =
+        service.postExtensionInfo(extensionInfo)
 
 }
