+++ domain/src/main/java/com/dsm/dms/domain/usecase/ApplyGoingOutUseCase.kt
@@ -2,17 +2,18 @@ package com.dsm.dms.domain.usecase
 
 import com.dsm.dms.domain.base.Result
 import com.dsm.dms.domain.base.UseCase
+import com.dsm.dms.domain.entity.GoingOutInfo
 import com.dsm.dms.domain.service.GoingOutService
 import io.reactivex.Single
 import io.reactivex.disposables.CompositeDisposable
 
 
-class DeleteGoingOutInfoUseCase(
+class ApplyGoingOutUseCase(
     private val service: GoingOutService,
     composite: CompositeDisposable
-): UseCase<Int, Result<Unit>>(composite) {
+): UseCase<GoingOutInfo, Result<Unit>>(composite) {
 
-    override fun create(id: Int): Single<Result<Unit>> =
-        service.deleteGoingOutInfo(id)
+    override fun create(goingOutInfo: GoingOutInfo): Single<Result<Unit>> =
+        service.postGoingOutInfo(goingOutInfo)
 
 }
