+++ domain/src/main/java/com/dsm/dms/domain/usecase/CancelGoingOutUseCase.kt
@@ -2,18 +2,17 @@ package com.dsm.dms.domain.usecase
 
 import com.dsm.dms.domain.base.Result
 import com.dsm.dms.domain.base.UseCase
-import com.dsm.dms.domain.entity.GoingOutInfo
 import com.dsm.dms.domain.service.GoingOutService
 import io.reactivex.Single
 import io.reactivex.disposables.CompositeDisposable
 
 
-class PostGoingOutInfoUseCase(
+class CancelGoingOutUseCase(
     private val service: GoingOutService,
     composite: CompositeDisposable
-): UseCase<GoingOutInfo, Result<Unit>>(composite) {
+): UseCase<Int, Result<Unit>>(composite) {
 
-    override fun create(goingOutInfo: GoingOutInfo): Single<Result<Unit>> =
-        service.postGoingOutInfo(goingOutInfo)
+    override fun create(id: Int): Single<Result<Unit>> =
+        service.deleteGoingOutInfo(id)
 
 }
